requirements.txt
Flask==2.3.2
torch==2.1.0
torchvision==0.15.2
transformers==4.30.0
Pillow==9.5.0

